refactor(evm-ext): Remove unused `AccountStorage`
This is a remnant of the EVM storage model that serialized the whole trie as a Move resource.
