+++ evm-ext/src/lib.rs
@@ -15,13 +15,13 @@ use {
     std::sync::LazyLock,
 };
 
+pub mod account;
 pub mod events;
 mod native_evm_context;
 mod native_impl;
 mod solidity_abi;
 mod state_changes;
 pub mod storage;
-pub mod trie_types;
 pub mod type_utils;
 
 /// Address where the EVM native is stored
