+++ evm-ext/src/type_utils.rs
@@ -11,7 +11,6 @@ use {
 };
 
 pub const ACCOUNT_INFO_PREFIX: &str = "Account_";
-pub const ACCOUNT_STORAGE_PREFIX: &str = "Storage_";
 
 pub fn account_info_struct_tag(address: &Address) -> StructTag {
     let name = format!("{ACCOUNT_INFO_PREFIX}{}", address.encode_hex());
